+++ QuoteSystem/Services/IAppDataRepository.cs
@@ -6,7 +6,7 @@ namespace QuoteSystem.Services
     {
 
         Task<IList<Quote>?> GetAllQuotesAsync();
-        Task<Quote?> GetQuoteAsync(int id);
+        Task<Quote?> GetQuoteAsync(long id);
 
         Task<bool> AddQuoteAsync(Quote quote);
         Task<bool> UpdateQuoteAsync(Quote quote);
